revert(airdate): reverts airdate offset & changes relative time to only display date (not time) (#1467)
* revert(airdate): reverts airdate offset and changes relative time to only display date (not time)

This reverts #1390 as it created more confusion when we offsetted the air date in relevance to the
timezone. It also changes the relative time to use date instead of time (so it will say `aired
yesterday` `today` `5 days ago` instead of `aired x hours ago` since we dont really the airtime
data.

* fix: relate time in days instead of hours

* fix: relative time in days

* fix: relative time in days (but properly)