fix: fixes pending media removal after request deletion
This fixes MORE TPYEORM LIFECYCLE HOOK ISSUES for postgresql. Current behavior is after media
request is removed (either by cancellation or deletion) the media remains as `pending` unless a user
with `manage requests` clears the media data, on **POSTGRESQL** database. With this change, the
media status will be changed to `UKNOWN` with the media request removal. TLDR; Though the
@AfterRemove() hook in MediaRequest works fine in SQLite it doesn't reliably update the media status
in PostgreSQL which is likely due to transaction timing or relation loading differences between the
two databases